Move AccountDetails to new module
