+++ src/components/TransactionConfirmationModal/index.tsx
@@ -15,7 +15,7 @@ import { isL2ChainId } from 'utils/chains'
 import Circle from '../../assets/images/blue-loader.svg'
 import { CloseIcon, CustomLightSpinner, ExternalLink } from 'theme'
 import { ExplorerDataType, getExplorerLink } from '../../utils/getExplorerLink'
-import { TransactionSummary } from '../AccountDetails/TransactionSummary'
+import { TransactionSummary } from '../../cow-react/modules/account/containers/AccountDetails/TransactionSummary'
 import { ButtonLight, ButtonPrimary } from '../Button'
 import { AutoColumn, ColumnCenter } from '../Column'
 import Modal from '@cow/common/pure/Modal'
