+++ src/cow-react/modules/account/containers/AccountDetails/Copy.tsx
