+++ src/cow-react/modules/account/containers/AccountDetails/SimpleAccountDetails.tsx
