+++ src/custom/components/EnhancedTransactionLink/index.tsx
@@ -1,5 +1,5 @@
 import { ExplorerLink } from 'components/ExplorerLink'
-import { GnosisSafeLink } from 'components/AccountDetails/Transaction/StatusDetails'
+import { GnosisSafeLink } from '@cow/modules/account/containers/Transaction/StatusDetails'
 
 import { EnhancedTransactionDetails, HashType } from 'state/enhancedTransactions/reducer'
 import { useWalletInfo } from '@cow/modules/wallet'
