+++ src/custom/components/Header/index.tsx
@@ -25,7 +25,7 @@ import {
 } from './styled'
 import MobileMenuIcon from './MobileMenuIcon'
 import { Web3Status } from '@cow/modules/wallet'
-import OrdersPanel from 'components/OrdersPanel'
+import { OrdersPanel } from '@cow/modules/account/containers/OrdersPanel'
 import NetworkSelector from 'components/Header/NetworkSelector'
 import CowBalanceButton from 'components/CowBalanceButton'
 import SVG from 'react-inlinesvg'
