+++ src/custom/components/TransactionConfirmationModal/index.tsx
@@ -15,14 +15,14 @@ import { CheckCircle, UserCheck } from 'react-feather'
 import GameIcon from 'assets/cow-swap/game.gif'
 import { Link } from 'react-router-dom'
 import { ConfirmationModalContent as ConfirmationModalContentMod, Wrapper } from './TransactionConfirmationModalMod'
-import { getStatusIcon } from 'components/AccountDetails'
+import { getStatusIcon } from '@cow/modules/account/containers/AccountDetails'
 import { OrderProgressBar } from 'components/OrderProgressBar'
 import { shortenAddress } from 'utils'
 import { getChainCurrencySymbols } from 'utils/gnosis_chain/hack'
 import { Routes } from '@cow/constants/routes'
 import { ActivityStatus, useMultipleActivityDescriptors } from 'hooks/useRecentActivity'
 import { getActivityState, useActivityDerivedState } from 'hooks/useActivityDerivedState'
-import { ActivityDerivedState } from 'components/AccountDetails/Transaction'
+import { ActivityDerivedState } from '@cow/modules/account/containers/Transaction'
 import AddToMetamask from '@cow/modules/wallet/api/components/AddToMetamask' // mod
 import { supportedChainId } from 'utils/supportedChainId'
 import { useOrder } from 'state/orders/hooks'
