+++ src/custom/hooks/useActivityDerivedState.ts
@@ -1,7 +1,7 @@
 import { SafeInfoResponse } from '@gnosis.pm/safe-service-client'
 import { useMemo } from 'react'
 import { getSafeWebUrl } from '@cow/api/gnosisSafe'
-import { ActivityDerivedState } from 'components/AccountDetails/Transaction'
+import { ActivityDerivedState } from '@cow/modules/account/containers/Transaction'
 import { EnhancedTransactionDetails } from 'state/enhancedTransactions/reducer'
 import { Order, OrderStatus } from 'state/orders/actions'
 import { getEtherscanLink } from 'utils'
